create SVM and 5G stocks
This study uses daily closing prices for two "5G-related" stocks from January 2019 to October 2020 and we analyze whether this historical data can help us predict price direction. We treat the problem of stock price forecasting as a binary classification problem and we use Support Vector Machine (SVM) as a learning technique to forecast stock prices direction from past data. The model attempts to predict whether a stock price in the future will be higher or lower than it is on a given day by predicting next day positive or negative move. A selection of suitable kernel functions and features are included before we run the model and we analyze its performance for hard margin and soft margin. We also evaluate the model using ROC, confusion matrix and other metrics. We tackle each task as required by the assignement and we also include along the report our results and observations related to classification accuracy for each method. We consider this analytical report a small but valuable contribution to the study and the development of prediction models for stock prices.